Split supervisor restart handoff helpers
